+++ backend/html_report/ui/src/components/distilled/metric/metric.css.ts
@@ -6,6 +6,9 @@ export default css`
     display: inline;
     
 }
-
+.rating {
+    width: 100%;
+    height: 100%;
+}
 
 `;